Fix #85
Previously we never reset `tries` after loading a new file, and since the
variable was an instance variable it would accumulate the total number of tries
to load any file, this caused issues when opening multiple files as these would
eventually reach the retry threshold and no new files would be able to be loaded.

See https://github.com/willprice/python-omxplayer-wrapper/issues/85 for discussion

Thanks to @jehutting for the bug hunting and fix.