Drop millibel volume settings and fix volume integration tests
Millibel volume setting is confusing so now we set it like so:

* 0 is muted
* 1 is normal volume
* 2 is twice as loud as normal volume

You can get to about 10 before you start clipping horrendously.

I also had to do a similar thing with `Mute` as we did with `Rate`: we
keep track of whether the player is muted and return the result of the
`Volume` property only if the player is not muted. If you call you get
the `Volume` property when the player is muted it just returns the
volume prior to muting