Better handle startup failures - do not leave dangling processes (#201)
* Move process killing to a separate method

This allows calling just the killing part of quit in a subsequent
commit. This only moves some code, without changing behaviour.

* Terminate process on startup errors

Before, if a process was successfully started, but there would be a
problem in subsequent setup (in particular the dbus connection could
fail), the process could keep running in the background.

This was particularly problematic when this caused the OMXPlayer
constructor to raise an exception, since then no player instance would
be created so the process pid would be lost.

This commit ensures that whenever the OMXPlayer constructor or load
method throws, no process will be dangling in the background.

This fixes #196

* Add more tests for player.quit

This adds test coverage for two exception handling cases.